CMakeLists: Fix import issues between C++ and QML
This took quite a bit of messing around with, but I'm much happier with how this is now. There were some code issues that were being hidden before that may have caused issues when importing the C++ module in the QML files. While it didn't seem to impact the actual program, it did hinder developing the project slightly.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>